Implement The Frontend and Backend Features
All frontend functionalities such as views for the project
including necessary javascript is implemented in this commit.
Some backend functionalities such as routing, necessary
controllers for CRUD operations for the customers, groups,
email scheduling, email template, migrations and database
set-up are also completed. The sending mail functions is not
implemented yet in this commit. Testing is also yet to be
carried out.